requirements.txt
flask
openai
numpy
scikit-learn
python-dotenv